backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.3
channels>=4.1
channels-redis>=4.2
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.4
django-cors-headers>=4.4
django-storages>=1.14
boto3>=1.35
cloudinary>=1.41
django-cloudinary-storage>=0.3
dj-database-url>=2.2
requests>=2.32
whitenoise>=6.7
daphne
